Select one or more files

Prompt the user to select one or more files (if allowed) of the given extensions using the standard system dialog.

Usage

openFileDialog(
  title = "Open File",
  filepath = getwd(),
  filters = "",
  description = "",
  multiple = F
)

Arguments

title

base::character Window title

filepath

base::character A default filename and/or path

filters

base::character File extension filter patterns

description

base::character Generic file type descriptor string

multiple

base::logical Allow multiple files to be selected?

Value

A character() vector of the selected file(s), or NULL if the dialog was cancelled.